backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
sqlalchemy[asyncio]>=2.0.30
asyncpg>=0.29.0
alembic>=1.13.1
redis>=5.0.4
celery>=5.4.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
pillow>=10.3.0
boto3>=1.34.96
httpx>=0.27.0
pydantic>=2.7.1
pydantic-settings>=2.2.1
python-dotenv>=1.0.1
aiofiles>=23.2.1
deepface>=0.0.93
opencv-python-headless>=4.9.0.80
tf-keras>=2.16.0
email-validator>=2.1.1
